exceptionHandler + refactor viewmodel creation
